made userid an int, which ended up touching a lot
Former-commit-id: 335c86dc7e2cb36ca6badbdefdefa5fab7547602 [formerly 60ca7913af5e236a6d015e3bd761629d75930d73]
Former-commit-id: 9a6c609a0fbd9287b830591b0c4cce202c36691c

+++ src/mitll/langtest/server/database/Database.java
@@ -24,13 +24,11 @@ public interface Database {
 
   void closeConnection(Connection connection);
 
-  void logEvent(String exid, String context, long userid, String device);
+  void logEvent(String exid, String context, int userid, String device);
 
   ServerProperties getServerProps();
 
   String getLanguage();
 
   LogAndNotify getLogAndNotify();
-
- // SessionManagement getSessionManagement();
 }
